+++ backend/src/main/java/com/project2/global/security/CustomAuthenticationSuccessHandler.java
@@ -1,7 +1,7 @@
 package com.project2.global.security;
 
 import com.project2.domain.member.entity.Member;
-import com.project2.domain.member.service.MemberService;
+import com.project2.domain.member.service.AuthService;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -18,7 +18,7 @@ import java.io.IOException;
 public class CustomAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
 
     private final Rq rq;
-    private final MemberService memberService;
+    private final AuthService authService;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
@@ -31,8 +31,8 @@ public class CustomAuthenticationSuccessHandler implements AuthenticationSuccess
         session.removeAttribute("redirectUrl");
 
         Member member = rq.getActor();
-        String accessToken = memberService.genAccessToken(member);
-        String refreshToken = memberService.genRefreshToken(member);
+        String accessToken = authService.genAccessToken(member);
+        String refreshToken = authService.genRefreshToken(member);
 
         rq.addCookie("accessToken", accessToken);
         rq.addCookie("refreshToken", refreshToken);
